pyproject.toml
[project]
name = "mca-mo-mobile"
version = "0.1.0"
description = "Audience remote to procedural art installation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "monsterui>=1.0.45",
    "python-dotenv>=1.2.2",
    "python-fasthtml>=0.13.3",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["."]
